Intercept all commands if rollkit.toml
