+++ cmd/rollkit/commands/run_node.go
@@ -6,11 +6,8 @@ import (
 	"math/rand"
 	"net/url"
 	"os"
-	"os/exec"
-	"path/filepath"
 	"time"
 
-	"github.com/BurntSushi/toml"
 	cmtcmd "github.com/cometbft/cometbft/cmd/cometbft/commands"
 	cometconf "github.com/cometbft/cometbft/config"
 	cometcli "github.com/cometbft/cometbft/libs/cli"
@@ -62,13 +59,7 @@ func NewRunNodeCmd() *cobra.Command {
 		Short:   "Run the rollkit node",
 		// PersistentPreRunE is used to parse the config and initial the config files
 		PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-			var err error
-			rollkitConfig, err = readTomlConfig()
-			if err != nil {
-				return err
-			}
-
-			err = parseConfig(rollkitConfig.Chain.ConfigDir, cmd)
+			err := parseConfig(cmd)
 			if err != nil {
 				return err
 			}
@@ -95,16 +86,6 @@ func NewRunNodeCmd() *cobra.Command {
 			return initFiles()
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if rollkitConfig.Entrypoint != "" {
-				// we need to add all user passed flags to the entrypoint command
-				// there is no better way to do this than read os.Args
-				flags := []string{}
-				if len(os.Args) >= 3 {
-					flags = os.Args[2:]
-				}
-				return runEntrypoint(rollkitConfig, "start", flags)
-			}
-
 			genDocProvider := cometnode.DefaultGenesisDocProviderFunc(config)
 			genDoc, err := genDocProvider()
 			if err != nil {
@@ -307,19 +288,17 @@ func initFiles() error {
 
 // parseConfig retrieves the default environment configuration, sets up the
 // Rollkit root and ensures that the root exists
-func parseConfig(configDir string, cmd *cobra.Command) error {
+func parseConfig(cmd *cobra.Command) error {
 	// Set the root directory for the config to the home directory
-	if configDir == "" {
-		configDir := os.Getenv("RKHOME")
-		if configDir == "" {
-			var err error
-			configDir, err = cmd.Flags().GetString(cometcli.HomeFlag)
-			if err != nil {
-				return err
-			}
+	home := os.Getenv("RKHOME")
+	if home == "" {
+		var err error
+		home, err = cmd.Flags().GetString(cometcli.HomeFlag)
+		if err != nil {
+			return err
 		}
 	}
-	config.RootDir = configDir
+	config.RootDir = home
 
 	// Validate the root directory
 	cometconf.EnsureRoot(config.RootDir)
@@ -328,85 +307,5 @@ func parseConfig(configDir string, cmd *cobra.Command) error {
 	if err := config.ValidateBasic(); err != nil {
 		return fmt.Errorf("error in config file: %w", err)
 	}
-	return nil
-}
-
-var ErrNoRollkitConfig = fmt.Errorf("no rollkit.toml found")
-
-func findConfigFile(startDir string) (string, error) {
-	dir := startDir
-	for {
-		configPath := filepath.Join(dir, "rollkit.toml")
-		if _, err := os.Stat(configPath); err == nil {
-			return configPath, nil
-		}
-
-		parentDir := filepath.Dir(dir)
-		if parentDir == dir {
-			break
-		}
-		dir = parentDir
-	}
-	return "", ErrNoRollkitConfig
-}
-
-func readTomlConfig() (rollconf.TomlConfig, error) {
-	var config rollconf.TomlConfig
-	startDir, err := os.Getwd()
-	if err != nil {
-		return config, fmt.Errorf("error getting current directory: %v", err)
-	}
-
-	configPath, err := findConfigFile(startDir)
-	if err != nil {
-		return config, err
-	}
-
-	if _, err := toml.DecodeFile(configPath, &config); err != nil {
-		return config, fmt.Errorf("error reading rollkit.toml: %v", err)
-	}
-
-	config.RootDir = filepath.Dir(configPath)
-
-	return config, nil
-}
-
-func runEntrypoint(rollkitConfig rollconf.TomlConfig, command string, args []string) error {
-	entrypointSourceFile := filepath.Join(rollkitConfig.RootDir, rollkitConfig.Entrypoint)
-	entrypointBinaryFile := filepath.Join(rollkitConfig.RootDir, rollupBinEntrypoint)
-
-	if !cometos.FileExists(entrypointBinaryFile) {
-		if !cometos.FileExists(entrypointSourceFile) {
-			return fmt.Errorf("rollkit: no such entrypoint file: %s", entrypointSourceFile)
-		}
-
-		// try to build the entrypoint
-		var buildArgs []string
-		buildArgs = append(buildArgs, "build")
-		buildArgs = append(buildArgs, "-o", entrypointBinaryFile)
-		buildArgs = append(buildArgs, entrypointSourceFile)
-		buildCmd := exec.Command("go", buildArgs...)
-		buildCmd.Stdout = os.Stdout
-		buildCmd.Stderr = os.Stderr
-		if err := buildCmd.Run(); err != nil {
-			return fmt.Errorf("rollkit: failed to build entrypoint: %w", err)
-		}
-	}
-
-	// run the entrypoint
-	var runArgs []string
-	runArgs = append(runArgs, command)
-	runArgs = append(runArgs, args...)
-	// we have to pass --home flag to the entrypoint, so it reads the correct config root directory
-	runArgs = append(runArgs, "--home", rollkitConfig.Chain.ConfigDir)
-	fmt.Printf("Running entrypoint: %s %v\n", entrypointBinaryFile, runArgs)
-	entrypointCmd := exec.Command(entrypointBinaryFile, runArgs...)
-	entrypointCmd.Stdout = os.Stdout
-	entrypointCmd.Stderr = os.Stderr
-
-	if err := entrypointCmd.Run(); err != nil {
-		return fmt.Errorf("rollkit: failed to run entrypoint: %w", err)
-	}
-
 	return nil
 }
